requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.2
python-docx>=0.8.11
streamlit>=1.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
